Update TenantScope::Decrypt to only contain a single CDO (#1902)
## Describe your changes

We have to build the ability to filter for tenant_roles that have a
certain permission (like the ability to decrypt an SSN). This query is
really difficult to write if `TenantScope::Decrypt` contains a list of
CDOs.

We already had the capability to support multiple `TS::Decrypt` for a
single role, so this change wasn't too invasive.

## Caveats

We currently serialize the list of scopes for a tenant role like 
```
[{"kind": "read"}, {"kind": "decrypt", "data": "full_address"}]
```
But now that all scope enum variants are simple, it makes me wonder if
we should serialize more closely to how we serialize DataIdentifiers,
like:
```
["read", "decrypt.full_address"]
```

This is a breaking change since these values are serialized in the DB.
But luckily, since we haven't started using tenant IAM much, there are
no decrypt perms stored yet :)

Same in dev and prod:

```
footprint=> select distinct scopes from tenant_role;
                          scopes
----------------------------------------------------------
 [{"kind": "read"}]
 [{"kind": "read"}, {"kind": "onboarding_configuration"}]
 [{"kind": "onboarding_configuration"}]
 [{"kind": "admin"}]
(4 rows)
```