Handle exceeding retries in doc upload (#1877)
## Describe your changes

Handle limiting the number of tries a user has. This PR makes it so a
user gets 3 vendor related failures before they are unable to move on.

I intentionally did not capture all failure cases (if we are the source
of the errors) since I think we can handle that separately. But could
add that in as well. I didn't have strong thoughts on introducing more
complexity

There's also the question of how to represent this in the document
request status. Since "failing due to retries" is a part of the overall
bifrost "Doc collection" step, and not an individual doc request itself.
I think in order to maintain a serialized log of why an entire set of
document requests failed, we'd need a new data model and this just
seemed simpler for now to encode in runtime logic


And as we discussed, there's no extra logic for trying selfie for a
subset of the times. You get 3 (or whatever) attempts and that's for
whatever the original doc request wanted you to collect. I don't think
the complexity is worth it at this point. if we feel it's too many
selfies, I'd rather just give ppl 2 attempts for both
## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

local testing