Allow decrypting selfies with TenantScope::DecryptDocument (#1986)
## Describe your changes

Resolves FP-2511

We decided to not control selfie decryption separately from docs. The
DataIdentifier is still separate for selfies, so we could decide to
enforce it separately in the future if we want.

But selfie is still the most strange DataIdentifier, so maybe we end up
removing it alltogether

## Testing
Checked in dev + prod that no role has this scope