Add asume endpoint for tenant rb auth (#2035)
## Describe your changes

Resolves FP-2573

Given an existing dashboard auth token, we check if the authed user is a
firm employee. If so, we mutate the session to instead contain a firm
employee session (rather than normal dashboard auth session) that gives
the firm employee access to read a given tenant_id's data.

## Caveats

One idea: should we catch any errors that occur in this endpoint at have
them 404 as if the path doesn't exist? Would make it nearly
imperceptible that we have an impersonation endpoint.

Not strictly necessary, but could be nice

## Test Plan

Will add integration tests in a future PR :)

For now:
[Screen Recording 2023-01-26 at
13.27.47.mov](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/f7b87633-ba12-476a-9a2d-bb5a533738e9/Screen%20Recording%202023-01-26%20at%2013.27.47.mov)