ID data validation: validate whole CDOs provided (#2045)
## Describe your changes

Part of
https://linear.app/footprint/issue/FP-2540/bifrost-address-validation-should-indicatehighlight-invalid-field

This adds brand new validation logic that isn't used anywhere. It is
mostly just scaffolding right now, and more functionality will be added
on later.

The intent is to be able to compose per-field error messages on which
fields weren't able to be validated. This PR starts by building the
scaffolding and adding validation that we don't provide any IDKs that
aren't part of a complete CDO. For example, you can't provide just
`id.first_name` or just `id.address_line1`

## Caveats

More functionality to validate each field coming in future PRs

## Test Plan

Unit tests!