Passkey Hint for identify on Bifrost (#2065)
Adds a new field to the Identify response that tells bifrost if any of
the pass keys on the account are synced: `has_syncable_pass_key`. This
means bifrost can opportunistically do Pass key auth on the
desktop/another device as long as the OS + browser supports it! More
cryptography and less SMS :)