Fix phone verification success msg (#2115
## Describe your changes

We had a flag to show the success message if both `identify` and `email`
requests were succeded. However, this last one just happens if the user
is new, so for any existing customers, we wouldn't display the
confirmation. Now the logic doesn't rely on the mutations, but rather in
a state (which is the only way to make this correct, as we need the
`identify` response)

Before:


https://user-images.githubusercontent.com/446731/216172835-21145a59-bae7-49d3-958e-07e978c2aba7.mov



Now - Existing user:


https://user-images.githubusercontent.com/446731/216172096-17c85376-2cec-4505-88a0-a0546275ec7b.mov


Now - New user:



https://user-images.githubusercontent.com/446731/216172886-34ae8ec3-6bf9-4a2a-ac4c-586d5fff8b71.mov


 

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->