[FP-2597] Fix members/roles pagination (#2079
## Describe your changes

- In the `users` page, we are using a cursor pagination. It has one big
downside: we can't build something like this (image below), because we
need to rely on the `next` wh

<img width="298" alt="Screenshot 2023-01-31 at 07 06 17"
src="https://user-images.githubusercontent.com/446731/215743727-6b5981e8-c314-430f-9406-4cf8ce93a9d8.png">

For the `members` and `roles` we built the backend pagination in a new
way, which will allow us to build a proper pagination component, like
the one above. So now we can send a `page` parameter to the endpoint -
much simpler: `?page=2`. However, for the `users`, we **still** use the
old style (cursor), because it has some complications in the backend.

Given we still have these two styles, what I did was separate the
pagination hook into two:
- `usePagination`
- `useCursorPagination`

Ideally, we should get rid of the cursor if possible. In the next PR,
I'll then create a proper pagination component, like the one above.

## Issue: FP-2597

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->