Rule engine v0.5 (#2122)
## Describe your changes

This PR attempts to add a more flexible way of adding RuleSets.

Concepts:

- Rule
    - function that take an input type T to a bool.
- RuleSet
- collections of functions (Rules) that take an input type T to a bool.
- RuleSetResult
    - the result of evaluating a RuleSet
- triggered
    - If any of the Rules in a RuleSet are `true`
- ActionableRuleSet
- A RuleSet that requires additional context for determining the
outcome. E.g. feature flag configuration or tenantID or other attribute
context
- ActionableRuleSetResult
    - a rule result for an ActionableRuleSet
- RuleSets (of either type, normal and Actionable) can do 2 things
(implement 2 traits)
    - EvaluateRuleSet (given an input of type T, produce an output)
- Determine if they have been triggered (given the output of evaluation,
tell me if the ruleset is triggered)
- This allows us to have multiple types of rules in the decision engine

In the purest form, `RuleSet`s are collections of functions (`Rule`s)
that take an input type `T` to a `bool`. The logic that is involved in
that function is arbitrarily complicated.

We want to use the rule sets in the decision engine to determine if a
user has passed AND/OR failed an onboarding. That is, it’s not enough
just to not fail an onboarding, they must affirmatively pass.

*However*, rulesets have no notion of what they are being used for, they
just execute the set of rules and return which rules were triggered or
not. Therefore, it is up the implementor to understand their purpose

Ruleset results have a notion of being `triggered` which is `true` if
any of the constituent rules in the ruleset are true.
- e.g. RulesetResult.triggered = {rule1 || rule2 || rule3 …}
- Since the logic inside each rule can be complex, you can make complex
&& conditions each individual rule itself, or a small ruleset.
-
## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

I wrote tests for DE and for rule set and tested locally with prod IDV
reqs and made sure the logs worked