Add tracing::instrument to all db util functions (#2171)
## Describe your changes

Just looked through all functions on models that take in a db connection
and traced them.

I don't love the `instrumented_connection.rs` that I added in #2103.
Will probably iterate on that strategy in a future PR. This adds a
different granularity of tracing that just helps to locate in which
function each DB query actually occurred