Add Csv<T> utils with Display implementation (#2184)
## Describe your changes

from conversation here:
https://github.com/onefootprint/monorepo/pull/2172#discussion_r1098902773

This bifurcates the Csv type into one for "output" (ie, serializing,
display, etc) vs "input" (deserializing). I tried adding this
functionality into a single Csv type, but it was quite difficult because
of the way that we deserialize.

The old `Csv` struct had a type constraint on `T` where `T:
DeserializeOwned`, which was required in order to have the derive of
`Deserialize`, which was required in order to use
`#[serde(deserialize_with = "deserialize_stringified_list")]`. This was
clunky if you had a type T that you wanted to Display as a
comma-separated list, but T wasn't dserializable.

This adds a new `newtypes::output::Csv` and uses it in many of our error
messages where we previously did stringification in custom ways