[FP-2732] Mark timeline events as portable when we portablize data (#2188)
## Describe your changes

When we mark data as portable after an approved onboarding decision, now
also make the user timeline events portable too.

## Caveats

We do deactivate some pieces of speculative data at onboarding approval
time IF AND ONLY IF there was another, more complete onboarding that was
started and finished for tenant B in between starting and finishing
onboarding for tenant A.

Not worth tracing down those user timeline events to leave them as
non-portable.

We also currently do the filtering of UserTimeline event kinds in RAM
since diesel makes it abundantly difficult to do a query on a jsonb
column. Maybe we'll reorganize this in the future and have the kind as a
separate column

## Test Plan

Tests!