Require checking perms before extracting tenant auth role/rolebinding (#2195)
## Describe your changes

We previously added a loophole to allow viewing the role of a tenant
auth class before `check_guard` was called. No need for this - can just
add the role to the TenantAuth trait.

Also added a `rolebinding()` function to TenantAuth that returns an
optional rolebinding. There aren't many use cases at all for it, and it
is one of the more weird ones since it is kind of TenantAuth
implementation specific... but, is useful in pr i will stack on top of
this :)