[FP-2750] Changes to avatar component (#2198)
Avatar cut company logos by applying a border-radius directly to the
image and making the image cover the whole surface. This works great for
images of people but not so well with logos. Since this component is
mainly meant to be used with company logos, we need to fix this.
<img width="88" alt="CleanShot 2023-02-08 at 10 05 20@2x"
src="https://user-images.githubusercontent.com/12160059/217585723-0c8cf4c8-8071-4dde-b21a-41fbe685d51e.png">