[FP-2771] Primitive permissions components + permissions for api keys views (#2219)
## Describe your changes

Creates two primitive helpers:

1) `usePermission`
Will return a `boolean` indicating if the user has or not permission to
a given scope:

```typescript
const { hasPermission } = usePermission();
hasPermission('api_keys') // true or false
```

2) `PermissionGate`

This component will check the permission for a scope. In case the user
doesn't have the scope, it will `disable` the element (in this case the
button) and show a tooltip with the fallback text

```typescript
<PermissionGate
	fallbackText={allT('permissions.api-keys.create')}
	scope="api_keys"
>
	<Button onClick={openCreateDialog} size="small" variant="secondary">
		{t('header.cta')}
	</Button>
</PermissionGate>
```

Demo:


<img width="404" alt="Screenshot 2023-02-09 at 16 03 07"
src="https://user-images.githubusercontent.com/446731/217913373-52ae81cb-cf53-4155-8356-7ae782497d35.png">
<img width="447" alt="Screenshot 2023-02-09 at 16 03 10"
src="https://user-images.githubusercontent.com/446731/217913381-cc78ae91-2727-4926-b53b-08356dd11ef9.png">


Also, it creates a helper to set the user in the session, such as
`asAdminUser`. Now we should take in account which user is logged, as
this will influence the tests.

## Issue: FP-2771

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->