Fix errors returned in doc enpoint (#2256)
## Describe your changes

When there's an issue that leads to an UploadFailed error, we create a
pending DocumentRequest

Currently the code is looking at the _current_ (i.e. pending) request
and not correctly surfacing errors.

This changes to look at the previous doc request so the frontend handles
correctly. Also changes to include the UploadFailed in the retry failure
logic

Fixes the handoff hanging issue @belcedogru and @keaganlong and I were
seeing


Also adds integration tests to test the logic of the GET API route

## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

Updated unit tests and local