Use stripe invoices instead of subscriptions (#2306)
## Describe your changes

This shifts the billing strategy significantly: before, we were using
stripe's subscriptions and interval-based prices on our products. KYC
checks were incremented (via a stripe API call) when they occurred, and
the quantity of users using PII storage per tenant was dumped to stripe
every time a new user signed up.

The primary benefit of using subscriptions is that they would
auto-generate invoices at the end of a billing period. We wouldn't need
any time-based job to run to fire off invoices.

But, there were some quirks with them: the billing interval for
non-metered prices showed charges for the next month's susbcription on
the invoice. Why? idk. It's not a big problem, but more the straw that
broke the camel's back: if an invoice is accidentally generated at the
wrong time from a subscription, or if we need to re-generate an invoice,
there wasn't really a great automated way to regenerate our invoice with
fixed logic.

*A normal invoice (how we generate after this PR):*


![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/c13285b1-98b6-48b3-82bc-d2e04e6baa01/image.png)

*The way stripe autogenerated invoices with subscription-based billing:*

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/95fad84c-82b8-4567-bdc1-6a35e29175a0/image.png)

Now, we moved invoice generation into this nice `POST /private/invoice`
endpoint that just sums up the quanity of KYC and PII vaulting to charge
for during this billing period. It creates a new draft invoice,
replacing the previous draft invoice if there was one.

So, the idea is that we can run this job on a daily (or whenever we
want) basis. We can always look at the amount owed to us by a tenant by
looking at their draft invoice. And at the end of the month, we can (for
now) manually graduate an invoice from a draft to a real invoice, which
will send the invoice to customers. This adds a nice manual check as we
start to roll out invoicing - maybe we will automate this in the future.

## Caveats

Now, we rely pretty heavily on these `select count(*)` queries in order
to determine how much to bill. An error in the query or forgetting to
update the query could be bad :) To solve this, will eventually probably
migrate to a system that emits events when a new charge is to be billed.
This job will sum up events by product to determine the quantities, and
we can compare our quantity to the `select count(*)` queries that we
currently do to audit the results. Then, our accounting system has to
fail in _two_ places to send out incorrect charges!

This still doesn't support billing for
- one-click onboardings
- reimbursing tenants when their users one-click onto other platforms
- charging for document scans
- charging for vault proxy

## Test Plan

stripe dashboard!

![image.png](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/e9b38bfe-565a-4254-bd5e-3f2cef6f7bf3/image.png)

Also resolves FP-2864