Clean up interface for adding data to UVW (#2400)
## Describe your changes

Update the API for adding data to the vault now that most of the
special-purpose interfaces (add_identity_data, add_email) are gone.

Also cleans up the interface for decomposing a `PutDataRequest`

## Caveats

Notably, we'll also now make one UserTimeline event for Email + other
identity data when the frontend starts sending email in the same req as
id data. Not a huge deal