[FP-3008] Use new API to create ob config for documents (#2476)
## Describe your changes

Instead of providing whether to collect docs/selfies with these custom
boolean flags, we take advantage of the work done in #2453 that
represents this information in CDOs.

## Caveats

The document/selfie information is sadly currently represented with
these booleans in the state machine for creating an ob config. In the
future, it could be represented in the frontend locally with CDOs - but
for now, we sadly just have some logic to map from these boolean flags
into the CDOs to send to the backend

## Test Plan

Tested locally