Replace PVDKind with IDK (#2524)
## Describe your changes

Since the VaultData table doesn't store email and phone number, we
created a separate enum, `PersonVaultDataKind`, to represent the subset
of IDK that could be stored in this table. It's getting a little clunky
to maintain, so let's just remove it. Added some checks at runtime in
the same place we would do the runtime conversion before, so we actually
don't lose any UX or safety.


In the future, we'll probably migrate Email + Phone to be stored in
VaultData anyways