Replace PVDKind with IDK (#2525)
## Describe your changes

This updates our old PvdBuilder, which would take in a
`DataRequest<IDK>` and save new VaultData rows, to also support
`DataRequest<BDK>`. Now, it's called VdBuilder and builds VaultData rows
from an IDK or BDK request.

Now, we'll actually write business vault data to the DB. But it's still
not fully functional - it won't deactivate any pieces of data on the
vault that already exist

## Issue: <FP-???>

## Caveats

Right now, the logic to add data to a business vault is conflated with
logic to add data to a personal vault. Will separate out in future PRs

## Test Plan

Will write tests shortly once this is much more fully functional :)