Calculate existing data for business vault wrapper (#2528)
## Describe your changes

Now that WriteableVw is parameterized on Business/Person vaults, it's
easy to calculate the list of populated business fields

## Caveats

This code isn't really used in business codepaths - it's used by the
VdBuilder to make sure you don't replace a "more full" CDO with a "less
full" CDO, like replacing ssn9 with ssn4. Since each business CD only
has one CDO, this is a no-op for all business data. There are definitely
some merits to not doing anything with this, but it's nice to follow the
pattern in case we do start supporting other CDO variants