[FP-3087] Create POST /entities/<>/vault/decrypt API (#2616)
## Describe your changes

This allows actually decrypting business data. We again create an
endpoint whose underlying code is shared with the user-specific
endpoint.

## Caveats

This is one of the last endpoints that I think we'll need to duplicate -
the others aren't as widely used and can be migrated to the new API that
works for both users and businesses. But I'm still thinking it would be
much nicer to have a simpler interface for pointing a second route to
the same handler

## Test Plan

Integration tests