[FP-3122] Add validation of investor profile data (#2673)
## Describe your changes

Adds enums to represent investor profile data and validates that the
data is parseable at validation time

## Caveats

Kind of annoying, but lists have to be provided as a string that is the
json serialized list. It might be worthwhile eventually having a
PiiJsonValue that is a scrubbed serde_json::Value

## Test Plan

Integration tests!