[FP-3202] Allow tenants to speculatively overwrite committed data of their users (#2744)
## Describe your changes

This allows fractional to write its own data to users that have
onboarded onto the platform.

The data that fractional writes to its own users' vaults will only be
visible to fractional since it is not committed.

## Caveats

There is still some undefined behavior that we'll probably have to
tackle in the future - what happens if
(1) user onboard onto tenant A
(2) tenant A updates info non-speculatively. This is scoped just to
tenant A
(3) user onboards onto tenant B, updates data, and this data becomes
portable

Should tenant A see the data from (2) or (3)? Right now, they will see
(2), but it might be more consistent behavior for them to see (3)

## Test Plan

Integration test