[FP-3227] Make ob config PK required in identify flow (#2753)
## Describe your changes

We used to support this flow without an ob config pk for making a my1fp
user, but it's proven difficult to support over time. Just rming this
for now, will probably redesign my1fp flows significantly when we launch
it.

For ex, we might want my1fp sign-up to use a version of bifrost with a
tenant PK similar to footprint live that has no permissions to decrypt

## Testing
Tested identify flow locally with user found and with new user