Migrate Email + PhoneNumber to VaultData/ContactInfo (#2748)
## Describe your changes

This moves Email and PhoneNumber e_data into the `VaultData` table,
where all other vaulted data lives.

We store auxiliary metadata info, like whether an email/phone is
verified, in the ContactInfo table, linked to the same lifetime as the
VaultData row and the Fingerprint.

A migration like this is much easier to do before we get a big customer
like Composer :)

## Caveats

This is very similar to the kv_data migration, but more risky... we
actually migrate the reads and writes to the new table in the same PR.
it is not backwards compatible, but given current volumes is worth
saving the complexity of a backwards compatible migration.

I leave all the old data around in the old PhoneNumber and Email tables
for now, so a code revert would bring us back to a working state (as
long as no new users sign up)

Sorry for the mondo PR :(

## Test Plan

Also tested that the email verification links that were sent out from a
previous version of code are backwards compatible - this is because we
keep the same PK from old Email rows when migrating to the ContactInfo
table


[Untitled.mov](https://graphite-user-uploaded-assets-prod.s3.amazonaws.com/xncBfoFfDL8XmXxkoXEI/5db46d75-4017-43dc-9492-31f824a144d2/Untitled.mov)