[FP-3234] Support changing orgs with the same auth token (#2811)
## Describe your changes

Previously, once you had selected an org to log into, you would have to
log out to log into a different org.

This now supports you using any tenant-user-based auth token to swap
orgs: a normal auth session, or a firm employee impersonation session.
And of course, we still support the "transient" workos auth session
which is just used after first authing to give the user a chance to
select the tenant.

On the frontend, when you're impersonating, we can now support a flow
where you go back to the org picker! And, if you belong to multiple
orgs, we can also support a very easy switcher

## Caveats

Also kind of crazy auth here. A cleaner way to do this doesn't
immediately come to mind.

## Test Plan

Tested locally:

Video shows first picking a tenant with a workos token, then going back
to the tenant picker to change to a different tenant, then impersonating
a tenant and going back to the tenant picker to change back to a tenant



https://user-images.githubusercontent.com/21066341/227401655-5a4c64dd-d53c-4f02-a1f9-ec229bfab51c.mov