[FP-3130] Add PUT entities/<>/vault api (#2823)
## Describe your changes

The last API to move over to the /entities/ path. We will likely open
this one up to the public API soon

## Caveats

Didn't end up making this a `/vault/<fp_id>/` path since it's not very
REST-y to provide the fp_id here... the vault is a part of the entity.
don't feel strongly though if we want to simplify the path

## Test Plan

This API isn't used by dashboard so doesn't need to be ported. Ported
integration tests though