Don't decrypt selfie if it doesn't exist for the document (#2837)
## Describe your changes

Voura bumped into this bug on the dashboard - if you don't have an ob config that requests selfie, but you try to decrypt just a document, you won't be able to since the frontend always requests to decrypt selfie. It's hard to solve on the frontend, and we're likely going to change the backend API soon so not worth jumping through frontend hoops

## Test Plan

Tested locally with an ob config that only requires the ID document. decryption of the doc in the dashboard succeeds after this change, but fails before it