WatchlistCheckTask error validation (#2822)
## Describe your changes

Adds 2 validation to WatchlistCheckTask:
- If onboarding.status =! Pass or the is_live=false, then we write a
WatchlistCheck with status=NotNeeded
- Went a bit back and forth as to whether we should even make a
WatchlistCheck in this situation but decided I liked it. I think we do
need to do this check in the task itself, and not just at task creation
(since we can schedule tasks in the future). I also think that instead
of just `tracing::` logging in this situation, it's nicer to leave a
strong papertrail by writing a WatchlistCheck row like this. (and
removes ambiguity in seeing a completed Task but no WC which could maybe
happen due to other issues)
- If the uservault does not contain the minimum data required for
Idology expectPA, then we write a WatchListCheck with status=Error

If we hit neither of those, we proceed to making vendor calls.


## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->