Add test that breaking changes aren't made to fingerprinting implementation (#2850)
## Describe your changes

I almost changed the implementation of IDK fingerprinting in a previous
PR to be very similar to the new implement of BDK (that uses the DI
string representation rather than IDK representation) but realized this
would cause breaking changes to all fingerprints that have been
generated in the past.

This test will just scream at you if you accidentally do that

## Test Plan

Tested by changing the implementation for IDK to use DIs, and the test
failed! yay