watchlist_check - make DI optional + add error (#2832)
## Describe your changes

Currently `watchlist_check` requires a foreign key to
`decision_intent_id`, with the idea being that every watchlist check
would have an associated vendor call. With the changes being made in
https://github.com/onefootprint/monorepo/pull/2822, if the watchlist
check is attempted but the vault does not have the sufficient data or
the vault's onboarding status is not Pass, then we will complete the
check without attempting vendor calls. This PR makes
`decision_intent_id` optional on `watchlist_check` so that we only write
this in cases where a vendor call was indeed attempted.

Also adds `error` to `watchlist_check` which will allow us to store an
enum specifying the reason for the watchlistcheck ending in an Error
status (currently just "RequiredDataNotPresent")


## Issue: <FP-???>

## Caveats

<!--- Are there any caveats to this changeset? -->

## Test Plan

<!--- How did you test your changes? :) -->