[FP-3362] Always decrypt business name in GET /entities (#2902)
## Describe your changes

For the business list page, we want to show the business name without
requiring an explicit decrypt. This decrypt business.name without making
an access log for every entity returned. If an entity doesn't have a
business.name (because it's a user vault or something), the decrypt
short circuits and we never talk to the enclave

This adds a `decrypted_attributes` (name suggestions welcome) map from
DI -> String to each entity that shows all of the "auto-decrypted"
attributes. This mirrors the response type of the normal decrypt
endpoint, so hopefully is easy to reuse some code on the frontend to
store the decrypted business.name

## Caveats
Right now, we have no batching of enclave reqs to decrypt business.name
- we just make 10 reqs in parallel. Probably an easy DOS attack
vector...

## Test Plan

Integration tests