Add CDO and DI for fully-KYCed BOs (#2956)
## Describe your changes

Replacing #2954 with this more fully-fledged approach that adds a
separate DI for fully-KYCed BOs. This makes a lot of downstream logic
easier because we don't have to pass an ob config around for validation,
it's just a function of the type of data identifier

Fully-KYCed BOs have different validation since we require that the
phone and email are provided. During the cleaning process, we will also
populate each record with a `link_id` that will eventually be used to
create a BO row in the DB

## Caveats

The link_id will be irrelevant for the primary BO... need to think about
a better way of making this not suck

## Test Plan

Tested w integration tests later in stack of PRs